+++ src/components/button/index.tsx
@@ -17,7 +17,7 @@ const styles = StyleSheet.create({
   // Button container
   button: {
     borderRadius: 50, // Rounded border
-    borderWidth: 2, // 2 point border widht
+    borderWidth: 2, // 2 point border width
     borderColor: '#FFFFFF', // White colored border
     paddingHorizontal: 50, // Horizontal padding
     paddingVertical: 10, // Vertical padding
